Refactor HomeView: Move each section to separate files
- Moved BannerSection to banner_section.dart
- Moved CategoriesSection to categories_section.dart
- Moved ProductsSection to products_section.dart
- Moved ProductsList to products_list.dart
- Updated HomeView to use the new separate section files
