+++ lib/features/home/presentation/widgets/categories_section/category_cilrcle_avatar.dart
