+++ lib/features/home/presentation/widgets/categories_section/shimer_for_category.dart
