+++ lib/features/home/presentation/widgets/products_section/card_item.dart
