+++ lib/features/home/presentation/widgets/products_section/shimmer_card_item.dart
